more ivms101 json serializers and tests
