+++ pkg/ivms101/ivms101.go
@@ -1,375 +1,477 @@
-/*
-Package ivms101 extends the Go protocol buffers generated by the ivms101 protobuf
-package with JSON loading utilities, validation helpers, short constants, etc.
-*/
 package ivms101
 
 import (
-	"strings"
-	"time"
+	"encoding/json"
+	"errors"
 )
 
-// Person converts a NaturalPerson into a Person protobuf message type.
-func (p *NaturalPerson) Person() *Person {
-	return &Person{
-		Person: &Person_NaturalPerson{
-			NaturalPerson: p,
-		},
-	}
+//
+// Person JSON
+//
+
+type serialPerson struct {
+	NaturalPerson *NaturalPerson `json:"naturalPerson,omitempty"`
+	LegalPerson   *LegalPerson   `json:"legalPerson,omitempty"`
 }
 
-// Person converts a LegalPerson into a Person protobuf message type.
-func (p *LegalPerson) Person() *Person {
-	return &Person{
-		Person: &Person_LegalPerson{
-			LegalPerson: p,
-		},
+func (p *Person) MarshalJSON() ([]byte, error) {
+	middle := serialPerson{
+		NaturalPerson: p.GetNaturalPerson(),
+		LegalPerson:   p.GetLegalPerson(),
 	}
+	return json.Marshal(middle)
 }
 
-// Validate the IVMS101 constraints for a natural person data definition. ON the first
-// invalid constraint found an error is returned.  No error is returned for valid data.
-func (p *NaturalPerson) Validate() (err error) {
-	// Constraint: required ValidNaturalPersonName
-	if p.Name == nil {
-		return ErrNoNaturalPersonNameIdentifiers
-	}
-	if err = p.Name.Validate(); err != nil {
+func (p *Person) UnmarshalJSON(data []byte) error {
+	middle := serialPerson{}
+	if err := json.Unmarshal(data, &middle); err != nil {
 		return err
 	}
 
-	// Constraint: ValidAddresses
-	for _, addr := range p.GeographicAddresses {
-		if err = addr.Validate(); err != nil {
-			return err
+	if middle.NaturalPerson != nil && middle.LegalPerson != nil {
+		return errors.New("person object cannot be both a natural and legal person")
+	}
+
+	i := Person{}
+	switch {
+	case middle.NaturalPerson != nil:
+		i.Person = &Person_NaturalPerson{
+			NaturalPerson: middle.NaturalPerson,
+		}
+	case middle.LegalPerson != nil:
+		i.Person = &Person_LegalPerson{
+			LegalPerson: middle.LegalPerson,
 		}
 	}
 
-	// Constraint: Optional ValidNationalIdentification
-	if p.NationalIdentification != nil {
-		if err = p.NationalIdentification.Validate(); err != nil {
-			return err
-		}
-	}
-
-	// Constraint: optional Max50Text
-	if len(p.CustomerIdentification) > 50 {
-		return ErrInvalidCustomerIdentification
-	}
-
-	// Constraint: Optional Valid DateAndPlaceOfBirth
-	if p.DateAndPlaceOfBirth != nil {
-		if err = p.DateAndPlaceOfBirth.Validate(); err != nil {
-			return err
-		}
-	}
-
-	// Constraint: Optional ISO-3166-1 alpha-2 codes or XX
-	if p.CountryOfResidence != "" {
-		// TODO: ensure the code is valid; for now just checking length
-		if len(p.CountryOfResidence) != 2 {
-			return ErrInvalidCountryCode
-		}
-
-		// Ensure that country code is all upper case
-		p.CountryOfResidence = strings.ToUpper(p.CountryOfResidence)
-	}
-
+	*p = i
 	return nil
 }
 
-// Validate the IVMS101 constraints for a legal person data definition. On the first
-// invalid constraint found an error is returned. No error is returned for valid data.
-func (p *LegalPerson) Validate() (err error) {
-	// Constraint: ValidLegalPersonName
-	// Constraint: LegalNamePresentLegalPerson
-	if p.Name == nil {
-		return ErrNoLegalPersonNameIdentifiers
+//
+// NaturalPerson JSON
+//
+
+type serialNaturalPerson struct {
+	Name               *NaturalPersonName      `json:"name,omitempty"`
+	Address            []*Address              `json:"geographicAddress,omitempty"`
+	Identification     *NationalIdentification `json:"nationalIdentification,omitempty"`
+	CustomerID         string                  `json:"customerIdentification,omitempty"`
+	DOB                *DateAndPlaceOfBirth    `json:"dateAndPlaceOfBirth,omitempty"`
+	CountryOfResidence string                  `json:"countryOfResidence,omitempty"`
+}
+
+func (n *NaturalPerson) MarshalJSON() ([]byte, error) {
+	middle := serialNaturalPerson{
+		Name:               n.Name,
+		Address:            n.GeographicAddresses,
+		Identification:     n.NationalIdentification,
+		CustomerID:         n.CustomerIdentification,
+		DOB:                n.DateAndPlaceOfBirth,
+		CountryOfResidence: n.CountryOfResidence,
 	}
-	if err = p.Name.Validate(); err != nil {
+	return json.Marshal(middle)
+}
+
+func (n *NaturalPerson) UnmarshalJSON(data []byte) error {
+	middle := serialNaturalPerson{}
+	if err := json.Unmarshal(data, &middle); err != nil {
 		return err
 	}
 
-	// Constraint: ValidAddresses
-	for _, addr := range p.GeographicAddresses {
-		if err = addr.Validate(); err != nil {
-			return err
-		}
+	i := NaturalPerson{
+		Name:                   middle.Name,
+		GeographicAddresses:    middle.Address,
+		NationalIdentification: middle.Identification,
+		CustomerIdentification: middle.CustomerID,
+		DateAndPlaceOfBirth:    middle.DOB,
+		CountryOfResidence:     middle.CountryOfResidence,
 	}
 
-	// Constraint: Optional Max50Text Datatype
-	if p.CustomerNumber != "" && len(p.CustomerNumber) > 50 {
-		return ErrInvalidCustomerNumber
-	}
-
-	// Constraint: Optional ValidNationalIdentification
-	if p.NationalIdentification != nil {
-		if err = p.NationalIdentification.Validate(); err != nil {
-			return err
-		}
-
-		// Constraint: ValidNationalIdentifierLegalPerson
-		if !(p.NationalIdentification.NationalIdentifierType == NationalIdentifierRAID ||
-			p.NationalIdentification.NationalIdentifierType == NationalIdentifierMISC ||
-			p.NationalIdentification.NationalIdentifierType == NationalIdentifierLEIX ||
-			p.NationalIdentification.NationalIdentifierType == NationalIdentifierTXID) {
-			return ErrValidNationalIdentifierLegalPerson
-		}
-
-		// Constraint: CompleteNationalIdentifierLegalPerson
-		// C9 means that Country of Issue must **only** be used for natural persons
-		if p.NationalIdentification.CountryOfIssue != "" {
-			return ErrCompleteNationalIdentifierLegalPerson
-		}
-		if p.NationalIdentification.NationalIdentifierType != NationalIdentifierLEIX {
-			// if the ID is not LEIX, Registration Authority is mandatory
-			if p.NationalIdentification.RegistrationAuthority == "" {
-				return ErrCompleteNationalIdentifierLegalPerson
-			}
-		} else {
-			// if the ID is an LEIX, Registration Authority must be empty
-			if p.NationalIdentification.RegistrationAuthority != "" {
-				return ErrCompleteNationalIdentifierLegalPerson
-			}
-		}
-	}
-
-	// Constraint: Optional ISO-3166-1 alpha-2 codes or XX
-	if p.CountryOfRegistration != "" {
-		// TODO: ensure the code is valid; for now just checking length
-		if len(p.CountryOfRegistration) != 2 {
-			return ErrInvalidCountryCode
-		}
-
-		// Ensure that country code is all upper case
-		p.CountryOfRegistration = strings.ToUpper(p.CountryOfRegistration)
-	}
-
+	// TODO warning: assignment copies lock value to *n
+	*n = i
 	return nil
 }
 
-// Validate the IVMS101 constraints for natural person name
-func (n *NaturalPersonName) Validate() (err error) {
-	// Constraint one or more
-	if len(n.NameIdentifiers) < 1 {
-		return ErrNoNaturalPersonNameIdentifiers
+//
+// NaturalPersonName and NaturalPersonNameIdentifiers JSON
+//
+
+type serialNaturalPersonName struct {
+	LocalNameIdentifiers    []*LocalNaturalPersonNameId `json:"localNameIdentifier,omitempty"`
+	NameIdentifiers         []*NaturalPersonNameId      `json:"nameIdentifier,omitempty"`
+	PhoneticNameIdentifiers []*LocalNaturalPersonNameId `json:"phoneticNameIdentifier,omitempty"`
+}
+
+func (n *NaturalPersonName) MarshalJSON() ([]byte, error) {
+	middle := serialNaturalPersonName{
+		LocalNameIdentifiers:    n.LocalNameIdentifiers,
+		NameIdentifiers:         n.NameIdentifiers,
+		PhoneticNameIdentifiers: n.PhoneticNameIdentifiers,
+	}
+	return json.Marshal(middle)
+}
+
+func (n *NaturalPersonName) UnmarshalJSON(data []byte) error {
+	middle := serialNaturalPersonName{}
+	if err := json.Unmarshal(data, &middle); err != nil {
+		return err
 	}
 
-	// Constraint: valid name identifiers
-	var legalNames int
-	for _, name := range n.NameIdentifiers {
-		if err = name.Validate(); err != nil {
-			return err
-		}
-
-		if name.NameIdentifierType == NaturalPersonLegal {
-			legalNames++
-		}
+	i := NaturalPersonName{
+		NameIdentifiers:         middle.NameIdentifiers,
+		LocalNameIdentifiers:    middle.LocalNameIdentifiers,
+		PhoneticNameIdentifiers: middle.PhoneticNameIdentifiers,
 	}
 
-	// Constraint: LegalNamePresent
-	if legalNames == 0 {
-		return ErrLegalNamesPresent
-	}
-
-	// Constraint: valid local name identifiers
-	for _, name := range n.LocalNameIdentifiers {
-		if err = name.Validate(); err != nil {
-			return err
-		}
-	}
-
-	// Constraint: valid phonetic name identifiers
-	for _, name := range n.PhoneticNameIdentifiers {
-		if err = name.Validate(); err != nil {
-			return err
-		}
-	}
-
+	// TODO warning: assignment copies lock value to *n
+	*n = i
 	return nil
 }
 
-// Validate the IVMS101 constraints for natural person name identifiers
-func (n *NaturalPersonNameId) Validate() (err error) {
-	if n.PrimaryIdentifier == "" || len(n.PrimaryIdentifier) > 100 {
-		return ErrInvalidNaturalPersonName
+type serialNaturalPersonNameId struct {
+	PrimaryIdentifier   string                    `json:"primaryIdentifier,omitempty"`
+	SecondaryIdentifier string                    `json:"secondaryIdentifier,omitempty"`
+	NameIdentifierType  NaturalPersonNameTypeCode `json:"nameIdentifierType,omitempty"`
+}
+
+func (p *NaturalPersonNameId) MarshalJSON() ([]byte, error) {
+	middle := serialNaturalPersonNameId{
+		PrimaryIdentifier:   p.PrimaryIdentifier,
+		SecondaryIdentifier: p.SecondaryIdentifier,
+		NameIdentifierType:  p.NameIdentifierType,
+	}
+	return json.Marshal(middle)
+}
+
+func (p *NaturalPersonNameId) UnmarshalJSON(data []byte) error {
+	middle := serialNaturalPersonNameId{}
+	if err := json.Unmarshal(data, &middle); err != nil {
+		return err
 	}
 
-	if len(n.SecondaryIdentifier) > 100 {
-		return ErrInvalidNaturalPersonName
+	i := NaturalPersonNameId{
+		PrimaryIdentifier:   middle.PrimaryIdentifier,
+		SecondaryIdentifier: middle.SecondaryIdentifier,
+		NameIdentifierType:  middle.NameIdentifierType,
 	}
 
-	typeCode := int32(n.NameIdentifierType)
-	if _, ok := NaturalPersonNameTypeCode_name[typeCode]; !ok {
-		return ErrInvalidNaturalPersonNameTypeCode
-	}
-
+	// TODO warning: assignment copies lock value to *p
+	*p = i
 	return nil
 }
 
-// Validate the IVMS101 constraints for local natural person anme identifiers
-func (n *LocalNaturalPersonNameId) Validate() (err error) {
-	if n.PrimaryIdentifier == "" || len(n.PrimaryIdentifier) > 100 {
-		return ErrInvalidNaturalPersonName
+type serialLocalNaturalPersonNameId struct {
+	PrimaryIdentifier   string                    `json:"primaryIdentifier,omitempty"`
+	SecondaryIdentifier string                    `json:"secondaryIdentifier,omitempty"`
+	NameIdentifierType  NaturalPersonNameTypeCode `json:"nameIdentifierType,omitempty"`
+}
+
+func (p *LocalNaturalPersonNameId) MarshalJSON() ([]byte, error) {
+	middle := serialLocalNaturalPersonNameId{
+		PrimaryIdentifier:   p.PrimaryIdentifier,
+		SecondaryIdentifier: p.SecondaryIdentifier,
+		NameIdentifierType:  p.NameIdentifierType,
+	}
+	return json.Marshal(middle)
+}
+
+func (p *LocalNaturalPersonNameId) UnmarshalJSON(data []byte) error {
+	middle := serialLocalNaturalPersonNameId{}
+	if err := json.Unmarshal(data, &middle); err != nil {
+		return err
 	}
 
-	if len(n.SecondaryIdentifier) > 100 {
-		return ErrInvalidNaturalPersonName
+	i := LocalNaturalPersonNameId{
+		PrimaryIdentifier:   middle.PrimaryIdentifier,
+		SecondaryIdentifier: middle.SecondaryIdentifier,
+		NameIdentifierType:  middle.NameIdentifierType,
 	}
 
-	typeCode := int32(n.NameIdentifierType)
-	if _, ok := NaturalPersonNameTypeCode_name[typeCode]; !ok {
-		return ErrInvalidNaturalPersonNameTypeCode
-	}
-
+	// TODO warning: assignment copies lock value to *p
+	*p = i
 	return nil
 }
 
-// Validate the IVMS101 constraints for legal person name.
-func (n *LegalPersonName) Validate() (err error) {
-	// Constraint: one or more
-	if len(n.NameIdentifiers) < 1 {
-		return ErrNoLegalPersonNameIdentifiers
+//
+// Address JSON
+//
+
+type serialAddress struct {
+	AddressType        AddressTypeCode `json:"addressType,omitempty"`
+	Department         string          `json:"department,omitempty"`
+	SubDepartment      string          `json:"subDepartment,omitempty"`
+	StreetName         string          `json:"streetName,omitempty"`
+	BuildingNumber     string          `json:"buildingNumber,omitempty"`
+	BuildingName       string          `json:"buildingName,omitempty"`
+	Floor              string          `json:"floor,omitempty"`
+	PostBox            string          `json:"postBox,omitempty"`
+	Room               string          `json:"room,omitempty"`
+	PostCode           string          `json:"postCode,omitempty"`
+	TownName           string          `json:"townName,omitempty"`
+	TownLocationName   string          `json:"townLocationName,omitempty"`
+	DistrictName       string          `json:"districtName,omitempty"`
+	CountrySubDivision string          `json:"countrySubDivision,omitempty"`
+	AddressLine        []string        `json:"addressLine,omitempty"`
+	Country            string          `json:"country,omitempty"`
+}
+
+func (a *Address) MarshalJSON() ([]byte, error) {
+	middle := serialAddress{
+		AddressType:        a.AddressType,
+		Department:         a.Department,
+		SubDepartment:      a.SubDepartment,
+		StreetName:         a.StreetName,
+		BuildingNumber:     a.BuildingNumber,
+		BuildingName:       a.BuildingName,
+		Floor:              a.Floor,
+		PostBox:            a.PostBox,
+		Room:               a.Room,
+		PostCode:           a.PostCode,
+		TownName:           a.TownName,
+		TownLocationName:   a.TownLocationName,
+		DistrictName:       a.DistrictName,
+		CountrySubDivision: a.CountrySubDivision,
+		AddressLine:        a.AddressLine,
+		Country:            a.Country,
+	}
+	return json.Marshal(middle)
+}
+
+func (a *Address) UnmarshalJSON(data []byte) error {
+	middle := serialAddress{}
+	if err := json.Unmarshal(data, &middle); err != nil {
+		return err
 	}
 
-	// Constraint: valid name identifiers
-	var legalNames int
-	for _, name := range n.NameIdentifiers {
-		if err = name.Validate(); err != nil {
-			return err
-		}
-
-		if name.LegalPersonNameIdentifierType == LegalPersonLegal {
-			legalNames++
-		}
+	i := Address{
+		AddressType:        middle.AddressType,
+		Department:         middle.Department,
+		SubDepartment:      middle.SubDepartment,
+		StreetName:         middle.StreetName,
+		BuildingNumber:     middle.BuildingNumber,
+		BuildingName:       middle.BuildingName,
+		Floor:              middle.Floor,
+		PostBox:            middle.PostBox,
+		Room:               middle.Room,
+		PostCode:           middle.PostCode,
+		TownName:           middle.TownName,
+		TownLocationName:   middle.TownLocationName,
+		DistrictName:       middle.DistrictName,
+		CountrySubDivision: middle.CountrySubDivision,
+		AddressLine:        middle.AddressLine,
+		Country:            middle.Country,
 	}
 
-	// Constraint: LegalNamePresent
-	if legalNames == 0 {
-		return ErrLegalNamesPresent
-	}
-
-	// Constraint: valid local name identifiers
-	for _, name := range n.LocalNameIdentifiers {
-		if err = name.Validate(); err != nil {
-			return err
-		}
-	}
-
-	// Constraint: valid phonetic name identifiers
-	for _, name := range n.PhoneticNameIdentifiers {
-		if err = name.Validate(); err != nil {
-			return err
-		}
-	}
-
+	// TODO warning: assignment copies lock value to *a
+	*a = i
 	return nil
 }
 
-// Validate the IVMS101 constraints for legal person name identifier
-func (n *LegalPersonNameId) Validate() (err error) {
-	if n.LegalPersonName == "" || len(n.LegalPersonName) > 100 {
-		return ErrInvalidLegalPersonName
+//
+// DateAndPlaceOfBirth JSON
+//
+
+type serialDateAndPlaceOfBirth struct {
+	DateOfBirth  string `json:"dateOfBirth,omitempty"`
+	PlaceOfBirth string `json:"placeOfBirth,omitempty"`
+}
+
+func (d *DateAndPlaceOfBirth) MarshalJSON() ([]byte, error) {
+	middle := serialDateAndPlaceOfBirth{
+		DateOfBirth:  d.DateOfBirth,
+		PlaceOfBirth: d.PlaceOfBirth,
+	}
+	return json.Marshal(middle)
+}
+
+func (d *DateAndPlaceOfBirth) UnmarshalJSON(data []byte) error {
+	middle := serialDateAndPlaceOfBirth{}
+	if err := json.Unmarshal(data, &middle); err != nil {
+		return err
 	}
 
-	typeCode := int32(n.LegalPersonNameIdentifierType)
-	if _, ok := LegalPersonNameTypeCode_name[typeCode]; !ok {
-		return ErrInvalidLegalPersonNameTypeCode
+	i := DateAndPlaceOfBirth{
+		DateOfBirth:  middle.DateOfBirth,
+		PlaceOfBirth: middle.PlaceOfBirth,
 	}
 
+	// TODO warning: assignment copies lock value to *d
+	*d = i
 	return nil
 }
 
-// Validate the IVMS101 constraints for local legal person name identifier
-func (n *LocalLegalPersonNameId) Validate() (err error) {
-	if n.LegalPersonName == "" || len(n.LegalPersonName) > 100 {
-		return ErrInvalidLegalPersonName
+//
+// NationalIdentification JSON
+//
+
+type serialNationalIdentification struct {
+	NationalIdentifier     string                     `json:"nationalIdentifier,omitempty"`
+	NationalIdentifierType NationalIdentifierTypeCode `json:"nationalIdentifierType,omitempty"`
+	CountryOfIssue         string                     `json:"countryOfIssue,omitempty"`
+	RegistrationAuthority  string                     `json:"registrationAuthority,omitempty"`
+}
+
+func (n *NationalIdentification) MarshalJSON() ([]byte, error) {
+	middle := serialNationalIdentification{
+		NationalIdentifier:     n.NationalIdentifier,
+		NationalIdentifierType: n.NationalIdentifierType,
+		CountryOfIssue:         n.CountryOfIssue,
+		RegistrationAuthority:  n.RegistrationAuthority,
+	}
+	return json.Marshal(middle)
+}
+
+func (n *NationalIdentification) UnmarshalJSON(data []byte) error {
+	middle := serialNationalIdentification{}
+	if err := json.Unmarshal(data, &middle); err != nil {
+		return err
 	}
 
-	typeCode := int32(n.LegalPersonNameIdentifierType)
-	if _, ok := LegalPersonNameTypeCode_name[typeCode]; !ok {
-		return ErrInvalidLegalPersonNameTypeCode
+	i := NationalIdentification{
+		NationalIdentifier:     middle.NationalIdentifier,
+		NationalIdentifierType: middle.NationalIdentifierType,
+		CountryOfIssue:         middle.CountryOfIssue,
+		RegistrationAuthority:  middle.RegistrationAuthority,
 	}
 
+	// TODO warning: assignment copies lock value to *n
+	*n = i
 	return nil
 }
 
-// Validate the IVMS101 constraints for a geographic address
-func (a *Address) Validate() (err error) {
-	// Constraint: valid required address type code
-	typeCode := int32(a.AddressType)
-	if _, ok := AddressTypeCode_name[typeCode]; !ok {
-		return ErrInvalidAddressTypeCode
+//
+// LegalPerson JSON
+//
+
+type serialLegalPerson struct {
+	Name                   *LegalPersonName        `json:"name,omitempty"`
+	Address                []*Address              `json:"geographicAddress,omitempty"`
+	CustomerNumber         string                  `json:"customerNumber,omitempty"`
+	NationalIdentification *NationalIdentification `json:"nationalIdentification,omitempty"`
+	CountryOfRegistration  string                  `json:"countryOfRegistration,omitempty"`
+}
+
+func (l *LegalPerson) MarshalJSON() ([]byte, error) {
+	middle := serialLegalPerson{
+		Name:                   l.Name,
+		Address:                l.GeographicAddresses,
+		CustomerNumber:         l.CustomerNumber,
+		NationalIdentification: l.NationalIdentification,
+		CountryOfRegistration:  l.CountryOfRegistration,
+	}
+	return json.Marshal(middle)
+}
+
+func (l *LegalPerson) UnmarshalJSON(data []byte) (err error) {
+	middle := serialLegalPerson{}
+	if err := json.Unmarshal(data, &middle); err != nil {
+		return err
 	}
 
-	// TODO: validate optional max length constraints
-	// Constraint: at most 7 address lines
-	if len(a.AddressLine) > 7 {
-		return ErrInvalidAddressLines
+	i := LegalPerson{
+		Name:                   middle.Name,
+		GeographicAddresses:    middle.Address,
+		CustomerNumber:         middle.CustomerNumber,
+		NationalIdentification: middle.NationalIdentification,
+		CountryOfRegistration:  middle.CountryOfRegistration,
 	}
 
-	// Constraint: ValidAddress
-	if len(a.AddressLine) == 0 && (a.StreetName == "" && (a.BuildingName == "" || a.BuildingNumber == "")) {
-		return ErrValidAddress
-	}
-
-	// Constraint: required valid country code
-	// TODO: validate ISO-3166-1 alpha-2 country code
-	if a.Country == "" || len(a.Country) != 2 {
-		return ErrInvalidCountryCode
-	}
-	a.Country = strings.ToUpper(a.Country)
-
+	// TODO warning: assignment copies lock value to *n
+	*l = i
 	return nil
 }
 
-// Validate the IVMS101 constraints for a national identification
-func (id *NationalIdentification) Validate() (err error) {
-	// TODO: Constraint ValidLEI
-	// Constraint: required Max35Text datatype
-	if id.NationalIdentifier == "" || len(id.NationalIdentifier) > 35 {
-		return ErrInvalidLEI
+//
+// LegalPersonName and LegalPersonNameIdentifiers JSON
+//
+
+type serialLegalPersonName struct {
+	LocalNameIdentifiers    []*LocalLegalPersonNameId `json:"localNameIdentifier,omitempty"`
+	NameIdentifiers         []*LegalPersonNameId      `json:"nameIdentifier,omitempty"`
+	PhoneticNameIdentifiers []*LocalLegalPersonNameId `json:"phoneticNameIdentifier,omitempty"`
+}
+
+func (l *LegalPersonName) MarshalJSON() ([]byte, error) {
+	middle := serialLegalPersonName{
+		LocalNameIdentifiers:    l.LocalNameIdentifiers,
+		NameIdentifiers:         l.NameIdentifiers,
+		PhoneticNameIdentifiers: l.PhoneticNameIdentifiers,
+	}
+	return json.Marshal(middle)
+}
+
+func (l *LegalPersonName) UnmarshalJSON(data []byte) error {
+	middle := serialLegalPersonName{}
+	if err := json.Unmarshal(data, &middle); err != nil {
+		return err
 	}
 
-	// Constraint: required valid national identifier type code
-	typeCode := int32(id.NationalIdentifierType)
-	if _, ok := NationalIdentifierTypeCode_name[typeCode]; !ok {
-		return ErrInvalidNationalIdentifierTypeCode
+	i := LegalPersonName{
+		NameIdentifiers:         middle.NameIdentifiers,
+		LocalNameIdentifiers:    middle.LocalNameIdentifiers,
+		PhoneticNameIdentifiers: middle.PhoneticNameIdentifiers,
 	}
 
-	// Constraint: valid country code
-	if id.CountryOfIssue != "" {
-		// TODO: validate ISO-3166-1 alpha-2 country code
-		if len(id.CountryOfIssue) != 2 {
-			return ErrInvalidCountryCode
-		}
-		id.CountryOfIssue = strings.ToUpper(id.CountryOfIssue)
-	}
-
-	// TODO: Contraint authority in GLEIF Registration authorities list
+	// TODO warning: assignment copies lock value to *l
+	*l = i
 	return nil
 }
 
-// Validate the IVMS101 constraints for date and place of birth
-func (d *DateAndPlaceOfBirth) Validate() (err error) {
-	// Constraint: require valid date
-	if d.DateOfBirth == "" {
-		return ErrInvalidDateOfBirth
+type serialLegalPersonNameId struct {
+	LegalPersonName               string                  `json:"legalPersonName,omitempty"`
+	LegalPersonNameIdentifierType LegalPersonNameTypeCode `json:"legalPersonNameIdentifierType,omitempty"`
+}
+
+func (p *LegalPersonNameId) MarshalJSON() ([]byte, error) {
+	middle := serialLegalPersonNameId{
+		LegalPersonName:               p.LegalPersonName,
+		LegalPersonNameIdentifierType: p.LegalPersonNameIdentifierType,
+	}
+	return json.Marshal(middle)
+}
+
+func (p *LegalPersonNameId) UnmarshalJSON(data []byte) error {
+	middle := serialLegalPersonNameId{}
+	if err := json.Unmarshal(data, &middle); err != nil {
+		return err
 	}
 
-	var date time.Time
-	if date, err = time.Parse("2006-01-02", d.DateOfBirth); err != nil {
-		return ErrInvalidDateOfBirth
+	i := LegalPersonNameId{
+		LegalPersonName:               middle.LegalPersonName,
+		LegalPersonNameIdentifierType: middle.LegalPersonNameIdentifierType,
 	}
 
-	if d.PlaceOfBirth == "" || len(d.PlaceOfBirth) > 70 {
-		return ErrInvalidPlaceOfBirth
+	// TODO warning: assignment copies lock value to *p
+	*p = i
+	return nil
+}
+
+type serialLocalLegalPersonNameId struct {
+	LegalPersonName               string                  `json:"legalPersonName,omitempty"`
+	LegalPersonNameIdentifierType LegalPersonNameTypeCode `json:"legalPersonNameIdentifierType,omitempty"`
+}
+
+func (p *LocalLegalPersonNameId) MarshalJSON() ([]byte, error) {
+	middle := serialLocalLegalPersonNameId{
+		LegalPersonName:               p.LegalPersonName,
+		LegalPersonNameIdentifierType: p.LegalPersonNameIdentifierType,
+	}
+	return json.Marshal(middle)
+}
+
+func (p *LocalLegalPersonNameId) UnmarshalJSON(data []byte) error {
+	middle := serialLocalLegalPersonNameId{}
+	if err := json.Unmarshal(data, &middle); err != nil {
+		return err
 	}
 
-	// Constraint: DateInPast
-	if date.After(time.Now()) {
-		return ErrDateInPast
+	i := LocalLegalPersonNameId{
+		LegalPersonName:               middle.LegalPersonName,
+		LegalPersonNameIdentifierType: middle.LegalPersonNameIdentifierType,
 	}
 
+	// TODO warning: assignment copies lock value to *p
+	*p = i
 	return nil
 }
