--- pkg/ivms101/testdata/legalperson.json
@@ -1,30 +1,0 @@
-{
-    "name": {
-        "name_identifiers": [
-            {
-                "legal_person_name": "Bob's Discount VASP, PLC",
-                "legal_person_name_identifier_type": 1
-            },
-            {
-                "legal_person_name": "Bob VASP",
-                "legal_person_name_identifier_type": 2
-            }
-        ]
-    },
-    "geographic_addresses": [
-        {
-            "address_type": 2,
-            "building_number": "762",
-            "street_name": "Grimsby Road",
-            "town_name": "Oxford",
-            "post_code": "OX8 U89",
-            "country": "GB"
-        }
-    ],
-    "customer_number": "",
-    "national_identification": {
-        "national_identifier": "213800AQUAUP6I215N33",
-        "national_identifier_type": 9
-    },
-    "country_of_registration": "GB"
-}
