+++ pkg/ivms101/testdata/national_identification.json
@@ -1,0 +1,6 @@
+{
+    "nationalIdentifier": "815026352",
+    "nationalIdentifierType": "DRLC",
+    "countryOfIssue": "TV",
+    "registrationAuthority": "RA777777"
+}
